% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionIntegrationApi.r
\name{ExpressionIntegrationApi_create_expression_group_sample_group_link}
\alias{ExpressionIntegrationApi_create_expression_group_sample_group_link}
\title{Create a link between a group of expression objects and a group of sample objects}
\usage{
ExpressionIntegrationApi_create_expression_group_sample_group_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the group of run-level objects (corresponding to a GCT file)}

\item{target_id}{The ID (accession) of the sample group object}
}
\description{
Create a link between a group of expression objects and a group of sample objects
}
