% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionIntegrationApi.r
\name{ExpressionIntegrationApi_create_expression_preparation_link}
\alias{ExpressionIntegrationApi_create_expression_preparation_link}
\title{Create a link between an expression object and a preparation}
\usage{
ExpressionIntegrationApi_create_expression_preparation_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the run-level object (corresponding to the column in a VCG/GCT file)}

\item{target_id}{The ID (accession) of the preparation object}
}
\description{
Create a link between an expression object and a preparation
}
