% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionIntegrationApi.r
\name{ExpressionIntegrationApi_delete_expression_library_link}
\alias{ExpressionIntegrationApi_delete_expression_library_link}
\title{Delete link between an expression object and a library}
\usage{
ExpressionIntegrationApi_delete_expression_library_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the run-level object (corresponding to the column in a VCG/GCT file)}

\item{target_id}{The ID (accession) of the library object}
}
\description{
Delete link between an expression object and a library
}
