% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionIntegrationApi.r
\name{ExpressionIntegrationApi_get_expression_by_preparation}
\alias{ExpressionIntegrationApi_get_expression_by_preparation}
\title{Retrieve expression run-level data by querying related preparation ID (accession)}
\usage{
ExpressionIntegrationApi_get_expression_by_preparation(
  id,
  page_limit,
  page_offset,
  use_versions,
  ...
)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{page_limit}{Maximum number of results to return. This value must be between 0 and 2000 (inclusive).}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}
}
\description{
Retrieve expression run-level data by querying related preparation ID (accession)
}
