% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionIntegrationApi.r
\name{ExpressionIntegrationApi_get_run_to_library_pairs}
\alias{ExpressionIntegrationApi_get_run_to_library_pairs}
\title{Retrieve run-library pairs by group id. Pagination is based on unique runs, not unique pairs.}
\usage{
ExpressionIntegrationApi_get_run_to_library_pairs(
  id,
  page_limit,
  page_offset,
  ...
)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{page_limit}{Maximum number of results to return. This value must be between 0 and 2000 (inclusive).}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}
}
\description{
Retrieve run-library pairs by group id. Pagination is based on unique runs, not unique pairs.
}
