% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometryIntegrationApi.r
\name{FlowCytometryIntegrationApi_create_flow_cytometry_group_sample_group_link}
\alias{FlowCytometryIntegrationApi_create_flow_cytometry_group_sample_group_link}
\title{Create a link between a group of flow-cytometry objects and a group of sample objects}
\usage{
FlowCytometryIntegrationApi_create_flow_cytometry_group_sample_group_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the group of run-level objects (corresponding to the column in a FACS file)}

\item{target_id}{The ID (accession) of the sample object}
}
\description{
Create a link between a group of flow-cytometry objects and a group of sample objects
}
