% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometryIntegrationApi.r
\name{FlowCytometryIntegrationApi_delete_flow_cytometry_sample_link}
\alias{FlowCytometryIntegrationApi_delete_flow_cytometry_sample_link}
\title{Delete link between a flow-cytometry object and a sample}
\usage{
FlowCytometryIntegrationApi_delete_flow_cytometry_sample_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the run-level data block}

\item{target_id}{The ID (accession) of the sample object}
}
\description{
Delete link between a flow-cytometry object and a sample
}
