% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometryIntegrationApi.r
\name{FlowCytometryIntegrationApi_get_flow_cytometry_by_sample}
\alias{FlowCytometryIntegrationApi_get_flow_cytometry_by_sample}
\title{Retrieve flow cytometry run-level data by querying related sample ID (accession)}
\usage{
FlowCytometryIntegrationApi_get_flow_cytometry_by_sample(
  id,
  page_limit,
  page_offset,
  use_versions,
  returned_metadata_fields,
  ...
)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{page_limit}{Maximum number of results to return. This value must be between 0 and 2000 (inclusive).}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}
}
\description{
Retrieve flow cytometry run-level data by querying related sample ID (accession)
}
