% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryIntegrationApi.r
\name{LibraryIntegrationApi_create_library_group_sample_group_link}
\alias{LibraryIntegrationApi_create_library_group_sample_group_link}
\title{Create links between samples and libraries}
\usage{
LibraryIntegrationApi_create_library_group_sample_group_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the library group object}

\item{target_id}{The ID (accession) of the sample group object}
}
\description{
Create links between samples and libraries
}
