% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryIntegrationApi.r
\name{LibraryIntegrationApi_delete_library_sample_link}
\alias{LibraryIntegrationApi_delete_library_sample_link}
\title{Delete a link between a library and a sample}
\usage{
LibraryIntegrationApi_delete_library_sample_link(source_id, target_id, ...)
}
\arguments{
\item{source_id}{The ID (accession) of the library object}

\item{target_id}{The ID (accession) of the sample object}
}
\description{
Delete a link between a library and a sample
}
