% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageApi.r
\name{LinkageApi_delete_link}
\alias{LinkageApi_delete_link}
\title{Deletes existing links matching the specified criteria.}
\usage{
LinkageApi_delete_link(first_id, second_id, first_type, second_type, ...)
}
\arguments{
\item{first_id}{Object ID (accession) (e.g. accession of study)}

\item{second_id}{Object ID (accession) (e.g. accession of study)}

\item{first_type}{Type of the object (e.g. study)}

\item{second_type}{Type of the object (e.g. study)}
}
\description{
Deletes existing links matching the specified criteria.
}
