% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageApi.r
\name{LinkageApi_save_links}
\alias{LinkageApi_save_links}
\title{Creates new links between objects.}
\usage{
LinkageApi_save_links(links, ...)
}
\arguments{
\item{links}{}
}
\description{
Creates new links between objects.
}
