% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsQueriesApi.r
\name{OmicsQueriesApi_search_flow_cytometry_groups}
\alias{OmicsQueriesApi_search_flow_cytometry_groups}
\title{Retrieve group objects by searching across multiple data types}
\usage{
OmicsQueriesApi_search_flow_cytometry_groups(
  study_filter,
  study_query,
  sample_filter,
  sample_query,
  search_specific_terms,
  vx_query,
  vx_filter,
  ex_query,
  ex_filter,
  fx_query,
  fx_filter,
  use_versions,
  cursor,
  page_limit,
  ...
)
}
\arguments{
\item{study_filter}{Filter by study metadata (key-value metadata pair(s)). E.g. \code{"Study Source"=ArrayExpress}}

\item{study_query}{Search for objects via a full-text query over all study metadata fields. E.g. \code{"RNA-Seq of human dendritic cells"}}

\item{sample_filter}{Filter by sample metadata (key-value metadata pair(s)). E.g. \code{"Species or strain"="Homo sapiens"}}

\item{sample_query}{Search for objects via a full-text query over all sample metadata fields. E.g. \code{Clozapine}}

\item{search_specific_terms}{If the full-text query term is present in an RTS dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders"  in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{vx_query}{Search for objects linked to variant data via data query (key-value pair(s)). E.g.  \code{VariationId=rs548419688|rs544419019}  \code{Intervals=12:23432-234324,30:234324-23432} \code{Gene=ASPM,BRCA1}  \code{Reference=A|T Alteration=C|G} \code{Quality=(0.9:1.0)}  \code{Type=SNP|MNP|INS|DEL|MIXED}  \code{info_VT=DEL}  \code{info_EUR_AF=(0.9:1.0)} \code{Start=12340} \code{AllelesNumber=1} \code{AlterationNumber=2}}

\item{vx_filter}{Filter by variant metadata (key-value metadata pair(s)). E.g. \code{"Variant Source"=dbSNP}}

\item{ex_query}{Search for objects linked to expression data via data query (key-value pair(s)). E.g. \code{Feature=ENSG00000230368,ENSG00000188976 MinValue=1.50}}

\item{ex_filter}{Filter by expression metadata (key-value metadata pair(s)). E.g. \code{"Genome Version"=hg19-BROAD}}

\item{fx_query}{Search for objects linked to flow cytometry data via data query (key-value pair(s)). E.g. \code{ReadoutType=Median|Count} \code{CellPopulation="CD45+, live"} \code{MinValue=3.5}}

\item{fx_filter}{Filter by flow cytometry metadata (key-value metadata pair(s)). E.g. \code{Organ=blood}}

\item{use_versions}{Specify which versions of omics data files are used in the query. By default the active version is used. See Versioning above. Syntax:  \\* or \code{v<version number>} or \code{<CHAIN_ID>}:\code{v<version number>} or \code{<CHAIN_ID>}:\code{<accession1,accession2,..>}}

\item{cursor}{The page tag to resume results from (see paging above).}

\item{page_limit}{How many results to retrieve per page. The default is 2000}
}
\description{
Retrieve group objects by searching across multiple data types
}
