% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsQueriesApi.r
\name{OmicsQueriesApi_search_streamed_expression_data}
\alias{OmicsQueriesApi_search_streamed_expression_data}
\title{Retrieve expression levels streamingly}
\usage{
OmicsQueriesApi_search_streamed_expression_data(
  group_accession,
  sample_filter,
  sample_query,
  search_specific_terms,
  feature_list,
  ...
)
}
\arguments{
\item{group_accession}{Accession of the group which contains the reference to the expression matrix}

\item{sample_filter}{Filter by sample metadata (key-value metadata pair(s)). E.g. \code{"Species or strain"="Homo sapiens"}}

\item{sample_query}{Search for objects via a full-text query over all sample metadata fields. E.g. \code{Clozapine}}

\item{search_specific_terms}{If the full-text query term is present in an RTS dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders"  in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{feature_list}{Filter results by specific feature (eg. Gene identifier). The feature parameter value must match the name of the identifier in the GCT file (under the NAME column). Example: \code{ENSG00000077044}}
}
\description{
Retrieve expression levels streamingly
}
