% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationIntegrationApi.r
\name{PreparationIntegrationApi_delete_preparation_group_sample_group_link}
\alias{PreparationIntegrationApi_delete_preparation_group_sample_group_link}
\title{Delete links between samples and preparations related to the specified group}
\usage{
PreparationIntegrationApi_delete_preparation_group_sample_group_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the preparation group object}

\item{target_id}{The ID (accession) of the sample group object}
}
\description{
Delete links between samples and preparations related to the specified group
}
