% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationIntegrationApi.r
\name{PreparationIntegrationApi_delete_preparation_sample_link}
\alias{PreparationIntegrationApi_delete_preparation_sample_link}
\title{Delete a link between a preparation and a sample}
\usage{
PreparationIntegrationApi_delete_preparation_sample_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the preparation object}

\item{target_id}{The ID (accession) of the sample object}
}
\description{
Delete a link between a preparation and a sample
}
