% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationIntegrationApi.r
\name{PreparationIntegrationApi_get_preparations_by_samples}
\alias{PreparationIntegrationApi_get_preparations_by_samples}
\title{Retrieve preparation metadata by querying related samples}
\usage{
PreparationIntegrationApi_get_preparations_by_samples(
  filter,
  query,
  search_specific_terms,
  page_limit,
  page_offset,
  returned_metadata_fields,
  ...
)
}
\arguments{
\item{filter}{Filter by sample metadata (key-value metadata pair(s)). E.g. \code{"Species or strain"="Homo sapiens"}}

\item{query}{Search for objects via a full-text query over all sample metadata fields. E.g. \code{Clozapine}}

\item{search_specific_terms}{If the full-text query term is present in an RTS dictionary, enabling this parameter will modify the query to include child terms of the full-text query.  For example, the search query "Body fluid" can be expanded to include the term "Blood" (a child term of  "Body fluid") so files containing either "Body fluid" or  "Blood"  in their metadata will be returned in the search results.  The parent-child relationship is defined by the key "broaders"  in the dictionary.  If the full query term is not present in a dictionary then this parameter has no effect.}

\item{page_limit}{Maximum number of results to return. This value must be between 0 and 2000 (inclusive).}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}
}
\description{
Retrieve preparation metadata by querying related samples
}
