% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleIntegrationApi.r
\name{SampleIntegrationApi_create_sample_group_study_link}
\alias{SampleIntegrationApi_create_sample_group_study_link}
\title{Create a link between a group of sample objects and a study}
\usage{
SampleIntegrationApi_create_sample_group_study_link(source_id, target_id, ...)
}
\arguments{
\item{source_id}{The ID (accession) of the sample group object}

\item{target_id}{The ID (accession) of the study object}
}
\description{
Create a link between a group of sample objects and a study
}
