% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleIntegrationApi.r
\name{SampleIntegrationApi_delete_sample_study_link}
\alias{SampleIntegrationApi_delete_sample_study_link}
\title{Delete link between a sample and a study}
\usage{
SampleIntegrationApi_delete_sample_study_link(source_id, target_id, ...)
}
\arguments{
\item{source_id}{The ID (accession) of the sample object}

\item{target_id}{The ID (accession) of the study object}
}
\description{
Delete link between a sample and a study
}
