% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyIntegrationApi.r
\name{StudyIntegrationApi_get_study_by_sample}
\alias{StudyIntegrationApi_get_study_by_sample}
\title{Retrieve study metadata by querying sample ID (accession)}
\usage{
StudyIntegrationApi_get_study_by_sample(id, returned_metadata_fields, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{returned_metadata_fields}{The parameter defines amount of metadata attributes to return:  1. \code{template} - return metadata attributes according to applied template, if object doesn’t have applied template default template will be used. This is the default for User endpoints. 2. \code{all} - return all metadata attributes with values and null attributes, if they are present in the applied template. This is the default for Curator endpoints.}
}
\description{
Retrieve study metadata by querying sample ID (accession)
}
