% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationSummaryApi.r
\name{ValidationSummaryApi_search_expression_data}
\alias{ValidationSummaryApi_search_expression_data}
\title{Retrieve validation summary by querying study ID (accession)}
\usage{
ValidationSummaryApi_search_expression_data(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve validation summary by querying study ID (accession)
}
