% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantIntegrationApi.r
\name{VariantIntegrationApi_create_variant_sample_link}
\alias{VariantIntegrationApi_create_variant_sample_link}
\title{Create a link between a variant object and a sample}
\usage{
VariantIntegrationApi_create_variant_sample_link(source_id, target_id, ...)
}
\arguments{
\item{source_id}{The ID (accession) of the run-level object (corresponding to the column in a VCG/GCT file)}

\item{target_id}{The ID (accession) of the sample object}
}
\description{
Create a link between a variant object and a sample
}
