% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantIntegrationApi.r
\name{VariantIntegrationApi_delete_variant_group_sample_group_link}
\alias{VariantIntegrationApi_delete_variant_group_sample_group_link}
\title{Delete link between a group of variant objects and a group of sample objects}
\usage{
VariantIntegrationApi_delete_variant_group_sample_group_link(
  source_id,
  target_id,
  ...
)
}
\arguments{
\item{source_id}{The ID (accession) of the group of run-level objects (corresponding to the column in a VCF file)}

\item{target_id}{The ID (accession) of the sample group object}
}
\description{
Delete link between a group of variant objects and a group of sample objects
}
