% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantIntegrationApi.r
\name{VariantIntegrationApi_get_run_to_sample_pairs}
\alias{VariantIntegrationApi_get_run_to_sample_pairs}
\title{Retrieve run-sample pairs by group id. Pagination is based on unique runs, not unique pairs.}
\usage{
VariantIntegrationApi_get_run_to_sample_pairs(id, page_limit, page_offset, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{page_limit}{Maximum number of results to return. This value must be between 0 and 2000 (inclusive).}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}
}
\description{
Retrieve run-sample pairs by group id. Pagination is based on unique runs, not unique pairs.
}
